Refactor microdata parsing
