+++ Sources/HtmlTinkerX/HtmlParserFromMicrodata.cs
@@ -26,7 +26,7 @@ public static class HtmlParserFromMicrodata {
         }
 
         IDocument document = HtmlParser.ParseWithAngleSharp(html);
-        return ExtractItems(document);
+        return MicrodataParser.ExtractItems(document);
     }
 
     /// <summary>
@@ -44,55 +44,4 @@ public static class HtmlParserFromMicrodata {
         return ParseMicrodataItems(content);
     }
 
-    private static List<HtmlMicrodataItem> ExtractItems(IDocument document) {
-        List<HtmlMicrodataItem> items = new();
-        foreach (var root in document.QuerySelectorAll("[itemscope]:not([itemprop])")) {
-            items.Add(ParseItem(root));
-        }
-        return items;
-    }
-
-    private static HtmlMicrodataItem ParseItem(IElement element) {
-        HtmlMicrodataItem item = new() {
-            Type = element.GetAttribute("itemtype"),
-            Id = element.GetAttribute("itemid")
-        };
-
-        foreach (var prop in element.QuerySelectorAll("[itemprop]")) {
-            if (!IsDirectChildOf(prop, element)) {
-                continue;
-            }
-            string name = prop.GetAttribute("itemprop") ?? string.Empty;
-            string value = GetPropertyValue(prop);
-            if (!item.Properties.TryGetValue(name, out var list)) {
-                list = new List<string>();
-                item.Properties[name] = list;
-            }
-            if (!string.IsNullOrEmpty(value)) {
-                list.Add(value);
-            }
-        }
-        return item;
-    }
-
-    private static bool IsDirectChildOf(IElement element, IElement parent) {
-        for (var node = element.ParentElement; node != null; node = node.ParentElement) {
-            if (node == parent) {
-                return true;
-            }
-            if (node.HasAttribute("itemscope") && !node.HasAttribute("itemprop")) {
-                break;
-            }
-        }
-        return false;
-    }
-
-    private static string GetPropertyValue(IElement element) {
-        return element.GetAttribute("content")
-            ?? element.GetAttribute("href")
-            ?? element.GetAttribute("src")
-            ?? element.GetAttribute("data")
-            ?? element.GetAttribute("value")
-            ?? element.TextContent.Trim();
-    }
 }
